+++ pkg/utils/kubernetes/compare/comparators.go
@@ -48,3 +48,8 @@ func ControlPlaneDeploymentOptionsDeepEqual(o1, o2 *operatorv1beta1.ControlPlane
 func NetworkOptionsDeepEqual(opts1, opts2 *operatorv1beta1.DataPlaneNetworkOptions) bool {
 	return reflect.DeepEqual(opts1.Services, opts2.Services)
 }
+
+// DataPlaneResourceOptionsDeepEqual checks if DataPlane resource options are equal.
+func DataPlaneResourceOptionsDeepEqual(opts1, opts2 *operatorv1beta1.DataPlaneResources) bool {
+	return reflect.DeepEqual(opts1, opts2)
+}